requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31
drf-yasg>=1.21
gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=2.1
